\name{get.beta.se.and.p}
\title{get.beta.se.and.p}
\alias{get.beta.se.and.p}
\description{A helper function for summarizing Nanostring results}
\usage{
get.beta.se.and.p(x, variable)
}
\arguments{
\item{x}{the response}
\item{variable}{the predictor}
}
\author{Daryl M. Waggott}
\keyword{Nanostring}
\keyword{Normalization}
\keyword{mRNA}  
\keyword{miRNA}
