\name{get.geo.mean}
\title{get.geo.mean}
\alias{get.geo.mean}
\description{A function to get the geometric mean.  Used in normalization functions}
\usage{
get.geo.mean(y)
}
\arguments{
  \item{y}{A vector of numeric values.  Values are assumed to be positive and unlogged.}
}
\value{
Returns the geometric mean.
}
\author{Daryl M. Waggott}
\keyword{NanoString}
\keyword{Normalization}
\keyword{mRNA}  
\keyword{miRNA}
