\name{NanoString}
\alias{NanoString}
\docType{data}
\title{
An rda file storing NanoString example datasets.  Currently, a data.frame consisting of Nanostring Rat RNA expression values.  The data were exported as a tab delimited file from the XLS results.
}
\description{
Contains mRNA data from an experiment of TCDD exposure in Rats.  The data structure and normalization workflow is different for miRNA.
}
\usage{data(NanoString)}
\format{A data.frame of 25 samples and 68 genes.}

\keyword{datasets}
