\name{NanoStringNorm-package}
\alias{NanoStringNorm-package}
\docType{package}
\title{A suite of functions for normalization, visualization and differential expression of NanoString miRNA and RNA expression levels}
\description{
A suite of functions for the normalization, visualization and differential expression of NanoString miRNA and RNA expression level.  The primary interface is through the function \emph{'NanoStringNorm'}.  The package enables existing preprocessing methods proposed by NanoString in addition to an extensible plug and play framework for further methods incorporation.  Emphasis is put on visualization of normalization diagnostics and results in order to prevent inadvertent batch effects or confounding introduced through the specifics of study design.
}
\details{
\tabular{ll}{
Package: \tab NanoStringNorm\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2011-09-15\cr
License: \tab GPL2 under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Author: Daryl M Waggott
Maintainer: Daryl Waggott <daryl.waggott@oicr.on.ca>
}
\keyword{ package }
