\name{NanoStringNorm-package}
\alias{NanoStringNorm-package}
\docType{package}
\title{A set of functions for Normalization of Nanostring miRNA and RNA expression levels.  Currently, only one function "NanoStringNorm" is used for all combinations of Normalization.}
\description{
A set of functions for Normalization of Nanostring miRNA and RNA expression levels.  Currently, only one function "NanoStringNorm" is used for all combinations of Normalization.
}
\details{
\tabular{ll}{
Package: \tab NanoStringNorm\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2011-09-15\cr
License: \tab GPL2 under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Author: Daryl M Waggott
Maintainer: Daryl Waggott <daryl.waggott@oicr.on.ca>
}
\keyword{ package }
